add plot view change;

+++ src/AvaloniaSqliteCurve/Entities/PointValue.cs
@@ -6,7 +6,7 @@
         
         public int PointId { get; set; }
         public double Value { get; set; }
-        public int Status { get; set; }
-        public long UpdateTime { get; set; }
+        public byte Status { get; set; }
+        public int UpdateTime { get; set; }
     }
 }
